Split script into randomKey() and main()
Function returns a random Key rather than a Scale. A Key can be
constructed given only the number of sharps; this makes it unnecessary
to keep a list of constant pitch-names, and in the future can allow the user to
set a range of accidentals to limit the keys printed.

From a Key one can also get the number of sharps, so that the number and
type of accidentals can be printed with each scale.