Add optional arguments to set number of accidentals and mode
The generated keys have less than or equal to the maximum number of
accidentals. By default this value is 7; it can be set using the -a
flag.

The mode (major or minor) of the generated keys can be set using the -M
(major) and the 'm (minor) flags. If not set, the mode of each key is
random.